# api/openapi.yaml
openapi: 3.1.0
info:
  title: GetLicense API
  version: 1.0.0
  description: |
    GetLicense is an open-source software licensing platform.
    It provides a REST API for managing products, licenses, machines,
    API keys, and webhooks with multi-tenant isolation and environment separation.

    ## Webhook Delivery

    Outbound webhook deliveries are at-least-once. Every POST sent to a
    customer endpoint includes a JSON envelope and a set of HMAC headers
    that let receivers authenticate the payload and detect duplicates.

    ### Envelope shape

    ```json
    {
      "id":          "<stable event id — equals domain_event_id>",
      "delivery_id": "<per-attempt webhook_event id>",
      "event_type":  "license.created",
      "data":        { ... },
      "timestamp":   "2026-04-24T12:34:56Z"
    }
    ```

    `id` is STABLE across duplicate fanouts (the at-least-once tail of a
    crash-before-checkpoint window) and across operator-initiated
    redeliveries of the same logical event. Receivers MUST dedupe by
    `id`. `delivery_id` is unique per attempt and is useful for log
    correlation when a single logical event is delivered multiple times.

    ### Headers

    The delivery POST sets three headers in addition to `Content-Type:
    application/json`:

    | Header | Meaning |
    | --- | --- |
    | `X-GetLicense-Timestamp` | Unix seconds at signing time (decimal) |
...
    | `X-GetLicense-Delivery-Id` | Per-attempt id (equal to `envelope.delivery_id`) |

...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      description: Exchanges a valid refresh token for a new access/refresh token pair.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/auth/logout-all:
    post:
      operationId: logoutAll
      tags: [auth]
      summary: Bulk-revoke every active session for the calling identity
      description: |
        Sets the per-identity session-invalidation cutoff to now: every JWT
        issued before this instant is rejected at verify time, regardless
        of jti. All refresh tokens for the identity are deleted in the
        same transaction. Requires identity authentication — API keys are
        unaffected.
      responses:
        "204":
          description: All sessions invalidated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/licenses/{id}:
    get:
      operationId: getLicense
      tags: [licenses]
      summary: Get a license
      description: Retrieves a single license by ID.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: License details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateLicense
      tags: [licenses]
      summary: Partial update of a license
      description: |
        Updates license.overrides or license.expires_at. Both are optional;
        omitted fields are unchanged. Sending overrides replaces the whole
        overrides struct. Sending expires_at=null clears it (makes the
        license perpetual).
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLicenseRequest"
      responses:
        "200":
          description: Updated license
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      operationId: revokeLicense
      tags: [licenses]
      summary: Revoke a license
      description: Permanently revokes a license. Only active or suspended licenses can be revoked.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: License revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/api-keys:
    post:
      operationId: createAPIKey
      tags: [api-keys]
      summary: Create an API key
      description: Creates a new API key for the authenticated account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAPIKeyResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listAPIKeys
      tags: [api-keys]
      summary: List API keys
      description: Returns a paginated list of API keys for the authenticated account.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Paginated list of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/api-keys/{id}:
    delete:
      operationId: deleteAPIKey
      tags: [api-keys]
      summary: Revoke an API key
      description: Revokes an API key by ID. Revoked keys can no longer authenticate.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: API key revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Webhooks ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Dual-mode authentication. Pass either a JWT access token (from login)
        or an API key (prefixed gl_live_ or gl_test_) as the Bearer token.
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      description: Page size (default 50, max 200)
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque cursor from a previous response's `next_cursor`. Omit for
        the first page. Treat as a black box — the encoding is not
        stable across API versions.
      schema:
        type: string
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, doc_url]
          properties:
            code:
              type: string
              description: Machine-readable error code
              examples: ["validation_error"]
            message:
              type: string
              description: Human-readable error description
              examples: ["Invalid product ID"]
            doc_url:
              type: string
              format: uri
              description: Link to error documentation
              examples: ["https://getlicense.io/docs/errors/validation_error"]

    # ── Pagination ───────────────────────────────────────────────────────
    # List endpoints use opaque cursor pagination. Clients should treat
    # next_cursor as a black box — it encodes the (created_at, id) pair
    # of the last row on the current page. Pass it back as the `cursor`
    # query parameter to fetch the next page.
    Page:
      type: object
      required: [data, has_more]
      properties:
        data:
          type: array
          description: Page of items; shape depends on the endpoint
        has_more:
          type: boolean
          description: True if additional pages exist after this one
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page; null when has_more is false

    # ── Account ──────────────────────────────────────────────────────────
...
    Identity:
      type: object
      required: [id, email, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        totp_enabled_at:
          type: string
          format: date-time
          nullable: true
          description: Set when TOTP two-factor is enabled, null otherwise.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── AccountSummary ────────────────────────────────────────────────────
    # Compact account view embedded inside membership, login, grant,
    # customer, and invitation responses. Use the full Account schema
    # when created_at is also needed.
    AccountSummary:
      type: object
      required: [id, name, slug]
      description: |
        Minimal public view of an account used to embed counterparty
        identity in Grant, Customer, and Invitation responses. Never
        includes created_at, member counts, or other account state.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string

    # ── MembershipSummary ────────────────────────────────────────────────
    # An identity's membership in an account along with its role,
    # denormalized for use in login/signup/me responses.
    MembershipSummary:
      type: object
      required: [membership_id, account, role_slug, role_name]
      properties:
        membership_id:
          type: string
          format: uuid
        account:
          $ref: "#/components/schemas/AccountSummary"
        role_slug:
          type: string
          description: Role slug (e.g. owner, admin, developer, operator).
        role_name:
          type: string
          description: Human-readable role name.

    # ── Role ─────────────────────────────────────────────────────────────
    # Permission bundle. Account-scoped (account_id set) for custom roles
    # or global preset (account_id null) for owner/admin/developer/operator.
...
    LicenseOverrides:
      type: object
      description: |
        Sparse per-license overrides for quantitative policy fields. Nil
        (omitted) fields mean "inherit from policy". Only quantitative
        fields are overridable — behavioral flags (Floating, Strict,
        ExpirationStrategy, RequireCheckout) are policy-only.
      properties:
        max_machines:
          type: integer
          nullable: true
        max_seats:
          type: integer
          nullable: true
        checkout_interval_sec:
          type: integer
          nullable: true
        max_checkout_duration_sec:
          type: integer
          nullable: true
        validation_ttl_sec:
          type: integer
          nullable: true
          minimum: 60
          maximum: 2592000
          description: |
            Per-license override of policy.validation_ttl_sec. Null inherits
            from policy. Same bounds as the policy field.
...
    License:
      type: object
      required:
        - id
        - account_id
        - product_id
        - policy_id
        - customer_id
        - key_prefix
        - token
        - status
        - environment
        - overrides
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
          description: Policy this license cascades from
        customer_id:
          type: string
          format: uuid
          description: End-user customer who owns this license
        key_prefix:
          type: string
          description: First characters of the license key for identification
        token:
          type: string
          description: Signed license token for offline verification
        status:
          type: string
          enum: [inactive, active, suspended, revoked, expired]
        overrides:
          $ref: "#/components/schemas/LicenseOverrides"
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Null for perpetual licenses; stamped on first activation if policy uses FROM_FIRST_ACTIVATION basis
        first_activated_at:
          type: string
          format: date-time
          nullable: true
        environment:
          $ref: "#/components/schemas/Environment"
        grant_id:
          type: string
          format: uuid
          nullable: true
          description: |
            Non-null when the license was created by a grantee exercising
            a LICENSE_CREATE capability under a grant. Identifies the
            grant whose constraints the license was issued under.
          examples: ["3b8fab1c-4b2e-4b1a-9f4e-4c6f6f8a9b10"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Machine ──────────────────────────────────────────────────────────
...
    Environment:
      type: string
      enum: [live, test]
...
    LoginResult:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, identity, memberships, current_account]
      properties:
        access_token:
          type: string
          description: JWT access token.
        refresh_token:
          type: string
          description: Opaque refresh token.
        token_type:
          type: string
          examples: ["Bearer"]
        expires_in:
          type: integer
          description: Access token TTL in seconds.
        identity:
          $ref: "#/components/schemas/Identity"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/MembershipSummary"
          description: All accounts this identity belongs to.
        current_account:
          $ref: "#/components/schemas/AccountSummary"
        needs_totp:
          type: boolean
          description: True when the identity has TOTP enabled and a second-factor step is required. When true, only pending_token is set.
        pending_token:
          type: string
          description: Short-lived token to submit with a TOTP code via POST /v1/auth/login/totp.
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    UpdateLicenseRequest:
      type: object
      description: |
        Partial update. Omitted fields are unchanged. For expires_at, the
        double-pointer distinguishes "unchanged" (omitted) from "explicitly null"
        (setting to null clears expiry, making the license perpetual).
      properties:
        overrides:
          $ref: "#/components/schemas/LicenseOverrides"
        expires_at:
          type: string
          format: date-time
          nullable: true
        customer_id:
          type: string
          format: uuid
          description: Reassign the license to a different customer in the same account.

    # ── Customer schemas (L4) ────────────────────────────────────────────
...
    CreateAPIKeyRequest:
      type: object
      additionalProperties: false
      required: [environment]
      properties:
        label:
          type: string
        environment:
          $ref: "#/components/schemas/Environment"
        scope:
          type: string
          enum: [account_wide, product]
          default: account_wide
          description: |
            Scope of the API key. account_wide grants access to all products
            in the account; product restricts the key to one product (set
            product_id below).
        product_id:
          type: string
          format: uuid
          nullable: true
          description: |
            Required when scope=product, MUST be omitted otherwise. Validated
            against the caller's tenant — unknown product_id returns 404.
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Optional expiration timestamp. Must be in the future.
        permissions:
          type: array
          items:
            type: string
          description: Optional per-key permission set.
        ip_allowlist:
          type: array
          items:
            type: string
            format: cidr
          description: Optional CIDR allowlist for source IPs.
    CreateAPIKeyResult:
      type: object
      required: [api_key, raw_key]
      properties:
        api_key:
          $ref: "#/components/schemas/APIKey"
        raw_key:
          type: string
          description: "Full API key value (shown only once)"

    # ── Webhook requests ─────────────────────────────────────────────────
...
    APIKeyList:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/APIKey"
...